data_engineering/sql_triggers/pyproject.toml
[tool.poetry]
name = "sql_triggers"
version = "0.1.0"
description = "An experiment showing the use of SQL triggers"
authors = ["Michal Bultrowicz <michal.bultrowicz@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
psycopg = "^3.0.15"
aiopg = "^1.3.4"
tenacity = "^8.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pylint = "^2.14.5"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
